backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
spacy==3.7.2
PyMuPDF==1.23.7
python-dotenv==1.0.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.1
python-docx==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1 